added ppm input format

+++ source/PixelData.h
@@ -106,7 +106,7 @@ public:
 
 	inline int GetWidth() const { return m_w; }
 	inline int GetHeight() const { return m_h; }
-	inline size_t GetPixelCount() { return m_w * m_h; }
+	inline int GetPixelCount() { return m_w * m_h; }
 	inline int GetComponentCount() const { return m_compCount; }
 
 };
